1.4 High-Quality Code/Exercise 3 Unit Testing - LinkedQueue/01.LinkedQueue/LinkedQueue.cs
namespace _01.LinkedQueue
{
    using System;
    using System.Collections;
    using System.Collections.Generic;
    using System.Linq;

    public class LinkedQueue<T> : IEnumerable<T>
    {
        private QueueNode<T> head;
        private QueueNode<T> tail;

        public int Count { get; private set; }

        public void Enqueue(T element)
        {
            var newNode = new QueueNode<T>(element);

            if (this.Count == 0)
            {
                this.tail = newNode;
                this.head = newNode;
            }
            else
            {
                this.tail.NextNode = newNode;
                this.tail = newNode;
            }

            this.Count++;
        }

        public T Dequeue()
        {
            if (this.Count == 0)
            {
                throw new InvalidOperationException("Queue is empty.");
            }

            var element = this.head.Value;
            this.head = this.head.NextNode;
            this.Count--;

            return element;
        }

        public bool Contains(T element)
        {
            return this.Any(el => el.Equals(element));
        }

        public void Clear()
        {
            this.head = null;
            this.tail = null;
            this.Count = 0;
        }

        public T Peek()
        {
            if (this.Count == 0)
            {
                throw new InvalidOperationException("Queue is empty.");
            }

            return this.head.Value;
        }

        public IEnumerator<T> GetEnumerator()
        {
            var currentEl = this.head;
            while (currentEl != null)
            {
                yield return currentEl.Value;
                currentEl = currentEl.NextNode;
            }
        }

        IEnumerator IEnumerable.GetEnumerator()
        {
            return this.GetEnumerator();
        }
    }
}
